+++ ForwardConfabulation.cpp
@@ -3,8 +3,8 @@
 ForwardConfabulation::ForwardConfabulation(size_t num_modules,
                                            const std::string &symbol_file,
                                            const std::string &master_file,
-                                           unsigned short min_single_occurrences,
-                                           unsigned short min_multi_occurrences)
+                                           uint8_t min_single_occurrences,
+                                           uint8_t min_multi_occurrences)
 {
     num_modules_ = num_modules;
 
@@ -20,13 +20,13 @@ ForwardConfabulation::ForwardConfabulation(size_t num_modules,
         }
     }
 
-    std::vector<unsigned short> level_sizes;
+    std::vector<uint8_t> level_sizes;
     level_sizes.push_back(num_modules_);
 
     Initialize(kb_specs, level_sizes, symbol_file, master_file, min_single_occurrences, min_multi_occurrences);
 }
 
-std::vector<std::string> ForwardConfabulation::Confabulation(const std::vector<std::string> &symbols, int index_to_complete, bool expectation)
+std::vector<std::string> ForwardConfabulation::Confabulation(const std::vector<std::string> &symbols, int8_t index_to_complete, bool expectation)
 {
     std::vector<std::string> result;
     if (!CheckArguments(symbols, index_to_complete)) {
@@ -34,14 +34,14 @@ std::vector<std::string> ForwardConfabulation::Confabulation(const std::vector<s
         return result;
     }
 
-    int index;
+    int8_t index;
     if (index_to_complete < 0) {
         index = AutoIndexToComplete(symbols);
     } else {
         index = index_to_complete;
     }
 
-    int actual_K = ActualK(symbols, index);
+    int32_t actual_K = ActualK(symbols, index);
     const std::unique_ptr<Module>& target_module = modules_[index];
     target_module->ExcitationsToZero();
 
@@ -60,13 +60,13 @@ std::vector<std::string> ForwardConfabulation::Confabulation(const std::vector<s
     return result;
 }
 
-int ForwardConfabulation::AutoIndexToComplete(const std::vector<std::string> &symbols)
+int8_t ForwardConfabulation::AutoIndexToComplete(const std::vector<std::string> &symbols)
 {
     if (symbols.empty() || symbols[0].empty()) {
         return -1;
     }
 
-    int index = FindFirstIndexOfSymbol(symbols, "");
+    int8_t index = FindFirstIndexOfSymbol(symbols, "");
 
     if (index >= 0) {
         return index;
@@ -76,7 +76,7 @@ int ForwardConfabulation::AutoIndexToComplete(const std::vector<std::string> &sy
     }
 }
 
-bool ForwardConfabulation::CheckIndex(const std::vector<std::string> &symbols, int index_to_complete)
+bool ForwardConfabulation::CheckIndex(const std::vector<std::string> &symbols, int8_t index_to_complete)
 {
     if (index_to_complete < 0) {
         std::cout << "Index to complete is negative" << "\n" << std::flush;
@@ -96,7 +96,7 @@ bool ForwardConfabulation::CheckIndex(const std::vector<std::string> &symbols, i
     return true;
 }
 
-bool ForwardConfabulation::CheckArguments(const std::vector<std::string> &symbols, int index_to_complete)
+bool ForwardConfabulation::CheckArguments(const std::vector<std::string> &symbols, int8_t index_to_complete)
 {
     if (symbols.empty()) {
         std::cout << "Symbols vector is empty" << "\n" << std::flush;
@@ -116,7 +116,7 @@ bool ForwardConfabulation::CheckArguments(const std::vector<std::string> &symbol
     }
 
     // autodetect mode
-    int index = AutoIndexToComplete(symbols);
+    int8_t index = AutoIndexToComplete(symbols);
 
     if (CheckIndex(symbols, index))
         return CheckVocabulary(symbols);
