+++ MultiLevelOrganizer.cpp
@@ -22,7 +22,7 @@
 #include "MultiLevelOrganizer.h"
 #include "utils/Utils.h"
 
-MultiLevelOrganizer::MultiLevelOrganizer(const std::vector<unsigned short>& level_sizes,
+MultiLevelOrganizer::MultiLevelOrganizer(const std::vector<uint8_t>& level_sizes,
                                          std::vector<std::unique_ptr<SymbolMapping>>&& level_mappings) :
     level_sizes_(level_sizes)
 {
