+++ TwoLevelSimpleConfabulation.h
@@ -9,14 +9,14 @@ public:
     TwoLevelSimpleConfabulation(size_t num_word_modules,
                                 const std::string &symbol_file,
                                 const std::string &master_file,
-                                unsigned short min_single_occurrences,
-                                unsigned short min_multi_occurrences);
+                                uint8_t min_single_occurrences,
+                                uint8_t min_multi_occurrences);
     TwoLevelSimpleConfabulation(const TwoLevelSimpleConfabulation& rhs) = delete;
     TwoLevelSimpleConfabulation& operator=(const TwoLevelSimpleConfabulation& rhs) = delete;
     TwoLevelSimpleConfabulation(TwoLevelSimpleConfabulation&& rhs) = delete;
     TwoLevelSimpleConfabulation&& operator=(TwoLevelSimpleConfabulation&& rhs) = delete;
 
-    virtual std::vector<std::string> Confabulation(const std::vector<std::string>& symbols, int index_to_complete, bool expectation);
+    virtual std::vector<std::string> Confabulation(const std::vector<std::string>& symbols, int8_t index_to_complete, bool expectation);
 
 private:
     size_t num_word_modules_;
